+++ src/main/ruby/truffleruby/core/time.rb
@@ -37,6 +37,10 @@
 class Time
   include Comparable
 
+  # Time#to_time is defined in date_core.c but it's just `return self`
+  # so we can make it available for use without having to `require "date"`.
+  alias_method :to_time, :itself
+
   def inspect
     str = strftime('%Y-%m-%d %H:%M:%S')
 
@@ -107,13 +111,15 @@ class Time
   def zone
     zone = Primitive.time_zone(self)
 
-    if zone && zone.ascii_only?
-      zone.encode Encoding::US_ASCII
-    elsif zone && Encoding.default_internal
-      zone.encode Encoding.default_internal
-    else
-      zone
+    if zone && Primitive.is_a?(zone, String)
+      if zone.ascii_only?
+        return zone.encode Encoding::US_ASCII
+      elsif Encoding.default_internal
+        return zone.encode Encoding.default_internal
+      end
     end
+
+    zone
   end
 
   # Random number for hash codes. Stops hashes for similar values in
@@ -166,10 +172,11 @@ class Time
   end
   alias_method :getutc, :getgm
 
-  def localtime(offset = nil)
+  def localtime(zone_or_offset = nil)
+    offset = zone_or_offset
     if offset
       to_utc = Time.send(:utc_offset_in_utc?, offset)
-      offset = Truffle::Type.coerce_to_utc_offset(offset)
+      offset = Truffle::Type.coerce_to_utc_offset(offset, self, :utc_to_local)
     end
 
     # the only cases when #localtime is allowed for a frozen time -
@@ -183,7 +190,9 @@ class Time
     if to_utc
       Primitive.time_utctime(self)
     else
-      Primitive.time_localtime(self, offset)
+      result = Primitive.time_localtime(self, offset)
+      Truffle::TimeOperations.set_zone_if_object(result, zone_or_offset)
+      result
     end
   end
 
@@ -347,8 +356,6 @@ class Time
     def at(sec, sub_sec = undefined, unit = undefined, **kwargs)
       # **kwargs is used here because 'in' is a ruby keyword
       timezone = kwargs[:in]
-      offset = timezone ? Truffle::Type.coerce_to_utc_offset(timezone) : nil
-      is_utc = utc_offset_in_utc?(timezone) if offset
 
       result = if Primitive.undefined?(sub_sec)
                  if Primitive.is_a?(sec, Time)
@@ -362,10 +369,15 @@ class Time
                    Primitive.time_at self, sec.to_i, ns
                  end
                end
+
+      offset = timezone ? Truffle::Type.coerce_to_utc_offset(timezone, result, :utc_to_local) : nil
+      is_utc = utc_offset_in_utc?(timezone) if offset
+
       if result && offset
         result = is_utc ? Primitive.time_utctime(result) : Primitive.time_localtime(result, offset)
       end
       if result
+        Truffle::TimeOperations.set_zone_if_object(result, timezone)
         return result
       end
 
@@ -426,9 +438,12 @@ class Time
         if utc_offset_in_utc?(utc_offset)
           utc_offset = :utc
         else
-          utc_offset = Truffle::Type.coerce_to_utc_offset(utc_offset)
+          zone = utc_offset
+          utc_offset = Truffle::Type.coerce_to_utc_offset(utc_offset, Time.utc(year, month, day, hour, minute, second), :local_to_utc)
         end
-        Truffle::TimeOperations.compose(self, utc_offset, year, month, day, hour, minute, second)
+        result = Truffle::TimeOperations.compose(self, utc_offset, year, month, day, hour, minute, second)
+        Truffle::TimeOperations.set_zone_if_object(result, zone)
+        result
       end
     end
 
@@ -442,7 +457,7 @@ class Time
       in_timezone = options[:in]
 
       if in_timezone
-        utc_offset = Truffle::Type.coerce_to_utc_offset(in_timezone)
+        utc_offset = Truffle::Type.coerce_to_utc_offset(in_timezone, time_now, :utc_to_local)
         is_utc = utc_offset_in_utc?(in_timezone)
         is_utc ? Primitive.time_utctime(time_now) : Primitive.time_localtime(time_now, utc_offset)
       else